# apps/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: ArtFi API
  version: 0.4.0
  description: Stage 4 Base Sepolia preview contract. Wallet intents never hold private keys; authenticated indexer writes are durable and replay-safe.
servers:
  - url: http://localhost:8080
paths:
...
  /v1/portfolio/{address}:
    get:
      operationId: getPortfolio
      parameters:
        - name: address
          in: path
          required: true
          schema:
            type: string
            pattern: "^0x[0-9a-fA-F]{40}$"
      responses:
        "200":
          description: Read-only public address portfolio
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Portfolio"
        "400":
          $ref: "#/components/responses/Problem"
  /v1/indexer/events:
    post:
      operationId: ingestChainEvent
      parameters:
        - name: X-Indexer-Key
          in: header
          required: true
          schema: { type: string, minLength: 16 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChainEventRequest"
      responses:
        "201": { description: Canonical event created }
        "200":
          {
            description: Replay,
            confirmation update,
            removal,
            or reorg replacement recorded,
          }
        "401": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
...
components:
...
  responses:
    Problem:
      description: RFC 9457-style request failure
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    Portfolio:
      type: object
      required:
        [
          address,
          chainId,
          network,
          positions,
          transactions,
          offers,
          notifications,
        ]
      properties:
        address: { type: string }
        chainId: { type: integer, const: 84532 }
        network: { type: string, const: base-sepolia }
        positions:
          type: array
          items:
            type: object
            required: [assetToken, symbol, balance, updatedAt]
            properties:
              assetToken: { type: string, pattern: "^0x[0-9a-fA-F]{40}$" }
              symbol: { type: string, maxLength: 12 }
              balance: { type: string, pattern: "^(0|[1-9][0-9]{0,77})$" }
              updatedAt: { type: string, format: date-time }
        transactions:
          type: array
          items: { type: object }
        offers:
          type: array
          items: { type: object }
        notifications:
          type: array
          items: { type: object }
    ChainEventRequest:
      type: object
      additionalProperties: false
      required:
        [
          chainId,
          transactionHash,
          logIndex,
          blockNumber,
          blockHash,
          contractAddress,
          eventName,
          payload,
          removed,
          confirmations,
        ]
      properties:
        chainId: { type: integer, const: 84532 }
        transactionHash: { type: string, pattern: "^0x[0-9a-fA-F]{64}$" }
        logIndex: { type: integer, minimum: 0 }
        blockNumber: { type: integer, minimum: 0 }
        blockHash: { type: string, pattern: "^0x[0-9a-fA-F]{64}$" }
        contractAddress: { type: string, pattern: "^0x[0-9a-fA-F]{40}$" }
        eventName: { type: string, minLength: 2, maxLength: 80 }
        payload: { type: object }
        removed: { type: boolean }
        confirmations: { type: integer, minimum: 0 }
...
    Problem:
      type: object
      required: [type, title, status, detail, requestId]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        requestId: { type: string }